# internal/api/api.yaml
openapi: 3.0.0
info:
  title: Payment Sandbox API
  version: 0.1.0
  description: Authentication APIs for merchant onboarding and login.
...
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://staging.api.example.com
    description: Staging server
  - url: https://api.example.com
    description: Production server
security:
  - bearerAuth: []
paths:
...
  /invoices:
    post:
      operationId: CreateInvoice
      summary: Create invoice
      tags: [Invoice]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceRequest"
      responses:
        "201":
          description: Invoice created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: ListInvoices
      summary: List invoices
      tags: [Invoice]
      parameters:
        - in: query
          name: merchantId
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
        - in: query
          name: page
          schema:
            type: string
        - in: query
          name: pageSize
          schema:
            type: string
      responses:
        "200":
          description: Invoice list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceListResponse"
        "500":
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/refresh:
    post:
      operationId: Refresh
      summary: Refresh JWT token
      tags: [Auth]
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid auth context
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateInvoiceRequest:
      type: object
      required: [merchantId, amount]
      properties:
        merchantId:
          type: string
        amount:
          type: integer
          format: int64
        currency:
          type: string
        dueDate:
          type: string
          format: date-time
...
    InvoiceResponse:
      type: object
      required: [id, invoiceNumber, merchantId, amount, currency, status, paymentToken, createdAt]
      properties:
        id:
          type: string
        invoiceNumber:
          type: string
        merchantId:
          type: string
        amount:
          type: integer
          format: int64
        currency:
          type: string
        status:
          type: string
        paymentToken:
          type: string
        dueDate:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
    InvoiceListResponse:
      type: object
      required: [items, page, size]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceResponse"
        page:
          type: integer
        size:
          type: integer
...
    AuthResponse:
      type: object
      required:
        - token
        - expiresAt
        - userId
        - email
        - role
      properties:
        token:
          type: string
        expiresAt:
          type: string
          format: date-time
        userId:
          type: string
        email:
          type: string
        role:
          type: string
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Detailed error message